set up database and run app
