+++ .gitignore
@@ -1,3 +1,5 @@
 .vscode
 __pycache__/
 *.pyc
+.env
+
