+++ .style.yapf
@@ -1,0 +1,3 @@
+[style]
+based_on_style = pep8
+COLUMN_LIMIT = 100
