+++ app/__init__.py
@@ -1,22 +1,21 @@
-# third party imports
-
-from flask import Flask
+from flask_api import FlaskAPI
 from flask_sqlalchemy import SQLAlchemy
 
-# local imports
-from config import app_config
+# local import
+from instance.config import app_config
 
-# db variable initialization
+# initialize sql-alchemy
 db = SQLAlchemy()
 
-
 def create_app(config_name):
-    app = Flask(__name__, instance_relative_config=True)
+    app = FlaskAPI(__name__, instance_relative_config=True)
     app.config.from_object(app_config[config_name])
     app.config.from_pyfile('config.py')
+    app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
     db.init_app(app)
 
     @app.route('/')
     def hello_world():
         return 'Hello, World!'
+
     return app
